# docs/swagger.yaml
basePath: /api
definitions:
  authn.LoginDto:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  authn.UserRegisterDto:
    properties:
      age:
        type: integer
      date_of_birth:
        type: string
      name_first:
        type: string
      name_last:
        type: string
      password:
        type: string
      place_of_birth:
        type: string
      username:
        type: string
    required:
    - age
    - date_of_birth
    - name_first
    - name_last
    - password
    - place_of_birth
    - username
    type: object
  task.CreateTaskDto:
    properties:
      description:
        type: string
      status:
        type: string
      title:
        type: string
      user_id:
        type: string
    required:
    - description
    - status
    - title
    - user_id
    type: object
  task.UpdateTaskDto:
    properties:
      description:
        type: string
      id:
        type: string
      status:
        type: string
      title:
        description: UserID      uuid.UUID `form:"user_id" binding:"required"`
        type: string
    required:
    - description
    - id
    - status
    - title
    type: object
info:
  contact: {}
paths:
  /authn/all-user:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: GetAlluser user
      tags:
      - authn
  /authn/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/authn.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - authn
  /authn/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: User credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/authn.UserRegisterDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register user
      tags:
      - authn
  /task/:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Task
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/task.UpdateTaskDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create Task
      tags:
      - task
    post:
      consumes:
      - application/json
      parameters:
      - description: Task
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/task.CreateTaskDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create Task
      tags:
      - task
  /task/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete Task
      tags:
      - task
    get:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create Task
      tags:
      - task
  /task/all:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create Task
      tags:
      - task
swagger: "2.0"